refactor: move setTags to a separate file
